test: Rewrote test to accurately represent bug
